Aug6 jen guide

+++ brewersproj/src/main/java/com/roomEase/brewersproj/controllers/ApplicationUserController.java
@@ -44,7 +44,7 @@ public class ApplicationUserController {
 
     // Mapping for signup form submission
     @PostMapping("/signup")
-    public RedirectView postSignup(String firstName, String lastName, String username, String password, String email, String householdId, String role, Integer telephone) {
+    public RedirectView postSignup(String firstName, String lastName, String username, String password, String email, String householdId, String role, Long telephone) {
         ApplicationUser user = new ApplicationUser();
         user.setFirstName(firstName);
         user.setLastName(lastName);
@@ -96,7 +96,9 @@ public class ApplicationUserController {
 
 
 
+
     // EVERYTHING above is working. You should be able to login and go to myprofiles page and create account
+
 
     // Mapping for user's profile page
     @GetMapping("/myprofile")
